refactor: Entity 설계 변경
* Driver 엔티티에서 공통 속성을 상위 클래스인 User 엔티티로 이동
