+++ BE/src/main/java/ifive/idrop/driver/dto/DriverSubscribeInfoResponse.java
@@ -44,7 +44,7 @@ public class DriverSubscribeInfoResponse {
                 .childName(child.getName())
                 .childBirth(child.getBirthDate())
                 .childGender(child.getGender().getLabel())
-                .childImage(child.getImage())
+                .childImage(child.getAvatar())
                 .startDate(startDate)
                 .endDate(endDate)
                 .startAddress(pickUpLocation.getStartAddress())
