+++ BE/src/main/java/ifive/idrop/dto/CurrentPickUpResponse.java
@@ -27,7 +27,7 @@ public class CurrentPickUpResponse {
         return CurrentPickUpResponse.builder()
                 .childId(pickUpInfo.getChild().getId())
                 .childName(pickUpInfo.getChild().getName())
-                .childImage(pickUpInfo.getChild().getImage())
+                .childImage(pickUpInfo.getChild().getAvatar())
                 .startDate(pickUpInfo.getPickUpSubscribe().getModifiedDate().plusDays(1))
                 .endDate(pickUpInfo.getPickUpSubscribe().getExpiredDate().minusDays(1))
                 .destination(Destination.of(pickUpInfo.getPickUpLocation()))
